requirements/dev.txt
#  Used for development environment requirements/dev.txt
# note common includes drivers for postgres, you may want to you sqllite
-r common.txt
django-debug-toolbar==0.9.4